+++ web/client-backplane/src/main/java/io/deephaven/javascript/proto/dhinternal/io/deephaven/proto/console_pb/figuredescriptor/ChartDescriptor.java
@@ -1,6 +1,3 @@
-/**
- * Copyright (c) 2016-2022 Deephaven Data Labs and Patent Pending
- */
 package io.deephaven.javascript.proto.dhinternal.io.deephaven.proto.console_pb.figuredescriptor;
 
 import elemental2.core.JsArray;
@@ -1716,7 +1713,7 @@ public class ChartDescriptor {
 
     public native int getColspan();
 
-    public native int getColumn();
+    public native double getColumn();
 
     public native boolean getIs3d();
 
@@ -1726,7 +1723,7 @@ public class ChartDescriptor {
 
     public native JsArray<MultiSeriesDescriptor> getMultiSeriesList();
 
-    public native int getRow();
+    public native double getRow();
 
     public native int getRowspan();
 
@@ -1755,7 +1752,7 @@ public class ChartDescriptor {
 
     public native void setColspan(int value);
 
-    public native void setColumn(int value);
+    public native void setColumn(double value);
 
     public native void setIs3d(boolean value);
 
@@ -1770,7 +1767,7 @@ public class ChartDescriptor {
         setMultiSeriesList(Js.<JsArray<MultiSeriesDescriptor>>uncheckedCast(value));
     }
 
-    public native void setRow(int value);
+    public native void setRow(double value);
 
     public native void setRowspan(int value);
