+++ web/client-backplane/src/main/java/io/deephaven/javascript/proto/dhinternal/io/deephaven/proto/console_pb/figuredescriptor/MultiSeriesSourceDescriptor.java
@@ -1,6 +1,3 @@
-/**
- * Copyright (c) 2016-2022 Deephaven Data Labs and Patent Pending
- */
 package io.deephaven.javascript.proto.dhinternal.io.deephaven.proto.console_pb.figuredescriptor;
 
 import elemental2.core.Uint8Array;
@@ -95,7 +92,7 @@ public class MultiSeriesSourceDescriptor {
 
     public native String getColumnName();
 
-    public native int getPartitionedTableId();
+    public native double getPartitionedTableId();
 
     public native int getType();
 
@@ -105,7 +102,7 @@ public class MultiSeriesSourceDescriptor {
 
     public native void setColumnName(String value);
 
-    public native void setPartitionedTableId(int value);
+    public native void setPartitionedTableId(double value);
 
     public native void setType(int value);
