+++ web/client-backplane/src/main/java/io/deephaven/javascript/proto/dhinternal/io/deephaven/proto/console_pb/figuredescriptor/chartdescriptor/ChartTypeMap.java
@@ -1,6 +1,3 @@
-/**
- * Copyright (c) 2016-2022 Deephaven Data Labs and Patent Pending
- */
 package io.deephaven.javascript.proto.dhinternal.io.deephaven.proto.console_pb.figuredescriptor.chartdescriptor;
 
 import jsinterop.annotations.JsOverlay;
@@ -33,7 +30,7 @@ public interface ChartTypeMap {
     int getPIE();
 
     @JsProperty(name = "TREEMAP")
-    int getTREEMAP();
+    double getTREEMAP();
 
     @JsProperty(name = "XY")
     int getXY();
@@ -54,7 +51,7 @@ public interface ChartTypeMap {
     void setPIE(int PIE);
 
     @JsProperty(name = "TREEMAP")
-    void setTREEMAP(int TREEMAP);
+    void setTREEMAP(double TREEMAP);
 
     @JsProperty(name = "XY")
     void setXY(int XY);
